# gen/openapiv2/api_v1.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "TitikTopup Core API",
    "description": "Core backend service for TitikTopup platform.",
    "version": "1.0.0"
  },
  "tags": [
    {
      "name": "TopupService"
    },
    {
      "name": "UserService"
    }
  ],
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/checkout": {
      "post": {
        "operationId": "TopupService_Checkout",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/titiktopupTransactionResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/titiktopupCheckoutRequest"
            }
          }
        ],
        "tags": [
          "TopupService"
        ]
      }
    },
    "/menu": {
      "get": {
        "operationId": "TopupService_GetMenu",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/titiktopupMenuResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "TopupService"
        ]
      }
    },
    "/user/profile": {
      "get": {
        "operationId": "UserService_GetProfile",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/titiktopupUserResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "UserService"
        ]
      }
    }
  },
  "definitions": {
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "titiktopupCategory": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "name": {
          "type": "string"
        },
        "slug": {
          "type": "string"
        },
        "products": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/titiktopupProduct"
          }
        }
      }
    },
    "titiktopupCheckoutRequest": {
      "type": "object",
      "properties": {
        "userIdGame": {
          "type": "string"
        },
        "zoneIdGame": {
          "type": "string"
        },
        "productId": {
          "type": "integer",
          "format": "int64"
        },
        "paymentMethod": {
          "type": "string"
        }
      }
    },
    "titiktopupMenuResponse": {
      "type": "object",
      "properties": {
        "categories": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/titiktopupCategory"
          }
        }
      }
    },
    "titiktopupProduct": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "sku": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "priceSell": {
          "type": "number",
          "format": "double"
        }
      }
    },
    "titiktopupTransactionResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "status": {
          "type": "string"
        },
        "totalPrice": {
          "type": "number",
          "format": "double"
        }
      }
    },
    "titiktopupUserResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "phone": {
          "type": "string"
        }
      }
    }
  }
}